jobs_project/requirements.txt
Scrapy
psycopg2-binary
pymongo # Optional
pandas